Optimize file CheckPermutation.py to be more Pythonic. Also renamed chapter folders to have the same naming style.
