--- Chapter 1/2_Check Permutation/CheckPermutation.py
@@ -1,35 +1,0 @@
-# O(N)
-import unittest
-from collections import Counter
-
-def check_permutation(string):
-    str1 = string[0]
-    str2 = string[1]
-    if len(str1) != len(str2):
-        return False
-    counter = Counter()
-    for c in str1:
-        counter[c] += 1
-    for c in str2:
-        if counter[c] == 0:
-            return False
-        counter[c] -= 1
-    return True
-
-class Test(unittest.TestCase):
-    dataT = [(['abcd', 'bacd']), (['3563476', '7334566']),
-             (['wef34f', 'wffe34'])]
-    dataF = [(['abcd', 'd2cba']), (['2354', '1234']), (['dcw4f', 'dcw5f'])]
-
-    def test_cp(self):
-        # true check
-        for test_string in self.dataT:
-            actual = check_permutation(test_string)
-            self.assertTrue(actual)
-        # false check
-        for test_string in self.dataF:
-            actual = check_permutation(test_string)
-            self.assertFalse(actual)
-
-if __name__ == "__main__":
-    unittest.main()
