+++ Chapter2/3_Delete_Middle_Node.py
