+++ Chapter2/4_Partition.py
