+++ Chapter2/5_Sum_Lists.py
