+++ Chapter2/6_Palindrome.py
