+++ Chapter2/7_Intersection.py
