+++ Chapter2/8_Loop_Detection.py
