+++ Chapter2/LinkedList.py
